// pavi-front/src/app/core/models/models.ts
export interface Empresa {
  codigo: number;
  razao_social: string;
  cnpj: string;
  cep: string;
  bairro: string;
  rua: string;
  numero: string;
  complemento: string;
  cod_cidade: number;
}